+++ api/python/src/MachO/init.cpp
@@ -58,6 +58,7 @@
 #include <LIEF/MachO/BuildVersion.hpp>
 #include <LIEF/MachO/FilesetCommand.hpp>
 #include <LIEF/MachO/ChainedBindingInfo.hpp>
+#include <LIEF/MachO/UnknownCommand.hpp>
 #include <LIEF/MachO/TwoLevelHints.hpp>
 #include <LIEF/MachO/LinkerOptHint.hpp>
 #include <LIEF/MachO/Builder.hpp>
@@ -111,6 +112,7 @@ void init_objects(nb::module_& m) {
   CREATE(ChainedBindingInfo, m);
   CREATE(TwoLevelHints, m);
   CREATE(LinkerOptHint, m);
+  CREATE(UnknownCommand, m);
   CREATE(Builder, m);
 }